Add a rpc server
This change introduces a remote procedure call server responsible for
handling requests to execute pre-registered procedures using JSON-RPC
2.0 as the transport protocol.

The purpose of this server is to provide a way for a mole process to
provide or be managed by others processes (e.g. some sort of user
interface that will creates/destroy instances of mole and/or collect
data from them).

The rpc server is disable by default. The user can enable it through the
`--rpc` flag. By default, the server will be listening on the localhost
address and on a random free port since a user might have multiple
instance of the program running at the same time on a single machine.
The `--rpc-address` flag can be use to set a specific address.